-Se creaton las BLL Amigos, Juegos, Prestamos. -Se creo el Contexto. -Se hicieron algunos ajustes en los registros, consultas y MainWindows

+++ Entidades/Juegos.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Text;
+
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace PrestamosdeJuegos.Entidades
 {
@@ -12,6 +14,6 @@ namespace PrestamosdeJuegos.Entidades
         public DateTime FechaCompra { get; set; } = DateTime.Now;
         public string Descripcion { get; set; }
         public double Precio { get; set; }
-        public int Existencia { get; set; }
+        public double Existencia { get; set; }
     }
 }
